Clicking middle buttons enqueues "previous" action.
Created new event case for every pane which handles mouse down? event. If the middle button was pressed it discards the event and enqueues "previous". If the event isn't discarded, the middle mouse acts as a left click which can unintentionally cause the window to close if the mouse is over certain BD elements. I would have wanted this behavior mapped to mouse 4 button press but that isn't captured by LV.