Fixes #23
Main window used to set it's FP window state to standard any time a VI's BD was inserted. Now it only sets tthe FP window state when it's current state is hidden (window isn't actually open). The panel close event also tracks the window state when it was closed so it opens up correctly (when a window is maximized it's standard state doesn't update to maximized size).

Everything other than Main.vi was updated as a result of MHL Date Cluster needing to be updated.