New Installation Sets Default Window Position
When the VIP installs it will check for teh vipeekWindowPosition token and set it to "50,50,1000,600" if not found. If you are upgrading the package it should respect the current peek window position but if it is a new install it will place the peek window in the upper left corner of the person's primary monitor with a size of 950x550 so it should never open up outside of someone's screen.